siuu

+++ Frontend/src/app/services/route-guard.service.ts
@@ -4,6 +4,7 @@ import { SnackbarService } from '../services/snackbar.service';
 import { AuthService } from './auth.service';
 import { jwtDecode } from "jwt-decode"
 import { GlobalConstants } from '../shared/global-constants';
+import { log } from 'console';
 @Injectable({
   providedIn: 'root'
 })
@@ -14,40 +15,61 @@ export class RouteGuardService {
 
   canActivate(router: ActivatedRouteSnapshot): boolean {
     let expectRoleArray = router.data;
-    expectRoleArray = expectRoleArray.expectRole;
+    expectRoleArray = expectRoleArray.expectedRole;
+
     const token: any = localStorage.getItem('accessToken');
 
     var tokenPayload: any;
 
     try {
       tokenPayload = jwtDecode(token);
+
     } catch (err) {
       localStorage.clear();
       this.router.navigate(['/']);
     }
 
-    let expectedRole = '';
-
-    for (let i = 0; i < expectRoleArray.length; i++) {
-      if (tokenPayload.role.some((r: any) => r == expectRoleArray[i])) {
-        expectedRole = tokenPayload.role;
-      }
-    }
-
-
-    if (tokenPayload.role == 'manager' || tokenPayload.role == 'employee') {
-      if (this.auth.isAuthenticated() && tokenPayload.role == expectedRole) {
-        return true;
-      }
-
-      this.snackbarService.openSnackBar(GlobalConstants.unauthroized, GlobalConstants.error);
-      this.router.navigate(['/cafe/dashboard']);
-      return false;
-    }
-    else {
+    if (!tokenPayload.role.includes('manager') && !tokenPayload.role.includes('employee')) {
       this.router.navigate(['/']);
       localStorage.clear();
       return false;
     }
+
+    let isAuthorized = false;
+    tokenPayload.role.some((r: any) => {
+      if (expectRoleArray.includes(r)) {
+        isAuthorized = true;
+      }
+    })
+
+    if (!isAuthorized) {
+      this.snackbarService.openSnackBar(GlobalConstants.unauthroized, GlobalConstants.error);
+      this.router.navigate(['/cafe/dashboard']);
+      return false;
+    }
+
+
+    // console.log(tokenPayload)
+    // console.log(expectedRole)
+
+    // for (let token of tokenPayload.role) {
+    //   if (token == 'manager' || token == 'employee') {
+    //     if (this.auth.isAuthenticated() && expectedRole.includes(token)) {
+    //       isAuthorized = true;
+    //       console.log(isAuthorized)
+    //       // return true;
+    //     }
+    //     // this.snackbarService.openSnackBar(GlobalConstants.unauthroized, GlobalConstants.error);
+    //     // this.router.navigate(['/cafe/dashboard']);
+    //     // return false;
+    //   }
+    //   else {
+    //     this.router.navigate(['/']);
+    //     localStorage.clear();
+    //     return false;
+    //   }
+    // }
+
+    return isAuthorized;
   }
 }
